ContactsPicker refactoring component
